user_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.1
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
aiofiles>=23.2.1
websockets>=12.0
pydantic>=2.5.2 